+++ ConfigManager.cs
@@ -11,6 +11,7 @@ namespace Batteries
         public static ConfigEntry<string> useBatteryKeybinds;
 
         public static ConfigEntry<float> batteryChargeAmount;
+        public static ConfigEntry<int> batteryMaxSpawns;
         public static ConfigEntry<int> batteryRarity;
         public static ConfigEntry<int> batteryScrapValue;
         public static ConfigEntry<int> batteryShopValue;
@@ -22,6 +23,7 @@ namespace Batteries
             useBatteryKeybinds = BatteriesMod.instance.Config.Bind("General Settings", "useBatteryKeybinds", DEFAULT_KEYBINDS, "The value for the keybinds to use a battery, split by comma. Values for keybinds can be found here: https://docs.unity3d.com/Packages/com.unity.inputsystem@1.0/api/UnityEngine.InputSystem.Key.html.");
 
             batteryChargeAmount = BatteriesMod.instance.Config.Bind("Battery Settings", "batteryChargeAmount", 0.25f, "The amount that batteries charge battery powered items. Can be set to a minimum of 0 (0% charge) and a maximum of 1 (100% charge).");
+            batteryMaxSpawns = BatteriesMod.instance.Config.Bind("Battery Settings", "batteryMaxSpawns", 10, "The maximum number of batteries that can spawn per moon. Can be set to any number greater than 0.");
             batteryRarity = BatteriesMod.instance.Config.Bind("Battery Settings", "batteryRarity", 70, "How rare it is for a battery to spawn. Can be set to a minimum of 0 (cannot spawn) and a maximum of 100 (common).");
             batteryScrapValue = BatteriesMod.instance.Config.Bind("Battery Settings", "batteryScrapValue", 5, "The scrap value of batteries.");
             batteryShopValue = BatteriesMod.instance.Config.Bind("Battery Settings", "batteryShopValue", 0, "How much batteries cost in the shop. When set to 0 or less batteries will not appear in the shop.");