+++ Plugin.cs
@@ -2,7 +2,6 @@
 using HarmonyLib;
 using LethalLib.Modules;
 using UnityEngine;
-using UnityEngine.InputSystem;
 
 namespace Batteries
 {
@@ -12,13 +11,14 @@ namespace Batteries
     {
         const string mGUID = "eXish.Batteries";
         const string mName = "Batteries";
-        const string mVersion = "1.0.5";
+        const string mVersion = "1.1.0";
 
         readonly Harmony harmony = new Harmony(mGUID);
 
         internal static BatteriesMod instance;
         internal static AssetBundle bundle;
         internal static AudioClip useBattery;
+        internal static Item batteryItem;
 
         void Awake()
         {
@@ -27,11 +27,13 @@ namespace Batteries
 
             ConfigManager.Init();
             if (string.IsNullOrEmpty(ConfigManager.useBatteryKeybinds.Value))
-                instance.Logger.LogWarning($"The value \"{ConfigManager.useBatteryKeybinds.Value}\" is not valid for setting \"useBatteryKeybind\"! The default will be used instead.");
+                instance.Logger.LogWarning($"The value \"{ConfigManager.useBatteryKeybinds.Value}\" is not valid for setting \"useBatteryKeybinds\"! The default will be used instead.");
             if (ConfigManager.batteryChargeAmount.Value < 0f || ConfigManager.batteryChargeAmount.Value > 1f)
                 instance.Logger.LogWarning($"The value \"{ConfigManager.batteryChargeAmount.Value}\" is not valid for setting \"batteryChargeAmount\"! The default will be used instead.");
             if (ConfigManager.batteryRarity.Value < 0 || ConfigManager.batteryRarity.Value > 100)
                 instance.Logger.LogWarning($"The value \"{ConfigManager.batteryRarity.Value}\" is not valid for setting \"batteryRarity\"! The default will be used instead.");
+            if (ConfigManager.batteryMaxSpawns.Value <= 0)
+                instance.Logger.LogWarning($"The value \"{ConfigManager.batteryMaxSpawns.Value}\" is not valid for setting \"batteryMaxSpawns\"! The default will be used instead.");
             if (ConfigManager.batteryScrapValue.Value < 0)
                 instance.Logger.LogWarning($"The value \"{ConfigManager.batteryScrapValue.Value}\" is not valid for setting \"batteryScrapValue\"! The default will be used instead.");
 
@@ -40,20 +42,22 @@ namespace Batteries
             if (bundle != null)
             {
                 useBattery = bundle.LoadAsset<AudioClip>("Assets/Batteries/UseBattery.wav");
-                Item battery = bundle.LoadAsset<Item>("Assets/Batteries/Battery.asset");
-                Utilities.FixMixerGroups(battery.spawnPrefab);
-                NetworkPrefabs.RegisterNetworkPrefab(battery.spawnPrefab);
-                if (ConfigManager.batteryRarity.Value < 0 || ConfigManager.batteryRarity.Value > 100)
-                    Items.RegisterScrap(battery, (int)ConfigManager.batteryRarity.DefaultValue, Levels.LevelTypes.All);
-                else if (ConfigManager.batteryRarity.Value != 0)
-                    Items.RegisterScrap(battery, ConfigManager.batteryRarity.Value, Levels.LevelTypes.All);
+                batteryItem = bundle.LoadAsset<Item>("Assets/Batteries/Battery.asset");
+                if (ConfigManager.batteryRarity.Value != 0)
+                {
+                    Utilities.FixMixerGroups(batteryItem.spawnPrefab);
+                    NetworkPrefabs.RegisterNetworkPrefab(batteryItem.spawnPrefab);
+                    Items.RegisterItem(batteryItem);
+                }
                 if (ConfigManager.batteryShopValue.Value > 0)
                 {
-                    UnityEngine.InputSystem.InputControlPath.ToHumanReadableString("<XRController>{RightHand}/gripButton", UnityEngine.InputSystem.InputControlPath.HumanReadableStringOptions.UseShortNames);
+                    Item storeBattery = bundle.LoadAsset<Item>("Assets/Batteries/StoreBattery.asset");
+                    Utilities.FixMixerGroups(storeBattery.spawnPrefab);
+                    NetworkPrefabs.RegisterNetworkPrefab(storeBattery.spawnPrefab);
                     TerminalNode node = ScriptableObject.CreateInstance<TerminalNode>();
                     node.clearPreviousText = true;
                     node.displayText = "Are you tired of having to travel all the way back to your ship to charge equipment? Well fear not! Durable's new battery line will solve all your charging needs! Get the power you deserve right when you need it!\n\nDurable is not responsible for any injury or death caused by our product. All items purchased are non-refundable.\n\n";
-                    Items.RegisterShopItem(battery, null, null, node, ConfigManager.batteryShopValue.Value);
+                    Items.RegisterShopItem(storeBattery, null, null, node, ConfigManager.batteryShopValue.Value);
                 }
             }
             else